Go: type MaxSizeBytes as *int64 to match other SDKs
Reviewer flagged that `*int` is architecture-dependent and narrower than the u64/long/Long used in the other SDKs. Switch to `*int64` (matches Java `Long` and .NET `long?`) and update the test literal cast.

+++ go/client_test.go
@@ -437,7 +437,7 @@ func TestSessionRequests_ReasoningSummary(t *testing.T) {
 func TestSessionRequests_PluginDirectoriesAndLargeOutput(t *testing.T) {
 	pluginDirs := []string{"/tmp/plugins/a", "/tmp/plugins/b"}
 	enabled := true
-	maxBytes := 1024
+	maxBytes := int64(1024)
 	largeOutput := &LargeToolOutputConfig{
 		Enabled:         &enabled,
 		MaxSizeBytes:    &maxBytes,
